rename getGeneratorNamespaces to lookupNamespaces and move to generate-blueprint
